model and spectrograms set to float

+++ main.py
@@ -94,7 +94,7 @@ for epoch in range(num_epochs):     # iterate over epochs
     for i, (spectrograms, labels) in enumerate(train_loader): # iterate over spectrograms and labels of train_loader
         print(spectrograms.shape)
         print(type(spectrograms))
-        spectrograms = spectrograms.double()
+        spectrograms = spectrograms.float()
         # send spectrogram and label data to selected device
         spectrograms = spectrograms.to(device)
         labels = labels.to(device)
